Create numberSequence
Bijection - a function between two sets, where each element of two sets is pired with
exactly one element of the other set

'*' is making the variable ready for the next task of the code 

mapcar - takes a list, call the function for the each element in the list seperately
creating a new list of results

lambda - function without a name

number-sequence - is making a range of numbers.

The function is setting p to 3, q to 5 and z-pq to a range of 0 to the multiple
 of p and q. Then mapcar calls the lambda function for the each element in the 
list z-pq.

Therefore, this function would be more in the functional programming style 
because ths function doesn't use loops. And loops are more recursive and take 
a lot of code instead of mapcar.